wxGUI/nviz: when showing legend parameter at can have floats not just int (merge from trunk, r68647)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_7_0@68649 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/nviz/mapwindow.py
@@ -447,7 +447,7 @@ class GLWindow(MapWindowBase, glcanvas.GLCanvas):
         if 0 in self.overlays:
             for param in self.overlays[0].cmd[1:]:
                 if param.startswith("at="):
-                    size = map(int, param.split("=")[-1].split(','))
+                    size = map(float, param.split("=")[-1].split(','))
                     break
         if size:
             wSize = self.GetClientSizeTuple()
